fix: moved loadResource functions to resource.js module
